huge refactor (started)
