+++ musictheatrebot.py
@@ -1,651 +1,15 @@
+#!/usr/bin/python
 from telegram.ext import Updater, CommandHandler
-import sched, time, random, logging, pickle, datetime, calendar
 from telegram.error import (TelegramError, Unauthorized, BadRequest, 
                             TimedOut, ChatMigrated, NetworkError)
-import datetime
-import numbers
-import gspread
-from oauth2client.service_account import ServiceAccountCredentials
-import numpy
-import re
 
-# spreadsheet
-columns = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K']
-wks = None
-def auth():
-    global wks
-    scope = ['https://spreadsheets.google.com/feeds',
-             'https://www.googleapis.com/auth/drive']
-    credentials = ServiceAccountCredentials.from_json_keyfile_name('credentials.json', scope)
-    gc = gspread.authorize(credentials)
-    wks = gc.open_by_key("1ExwdtbLUBpWZ12fg2faURtZLf7T8VZa0tndEX4SYkck").get_worksheet(0)
+import logging
 
-configFile = "./session.pk"
-watb = -1001049406492
-newseeds = -1001138132564
-retardStickerId = "CAADBAAD2wADeyqRC60Pvd---1a5Ag";
+# add files
+import core, session
+import other_commands, start_commands, session_commands, sheet_commands
 
-admins = [
-          "Xanes",
-          "ysoftware",
-          "tbshfmn",
-          "sexy_nutella_69",
-          "amobishoproden",
-          "jntn7",
-          "Doomgoat",
-          "Tom_veldhuis",
-          "FkinTag"
-          ]
-
-# utilities
-
-fValue = lambda cell: cell.value
-fNonEmpty = lambda value: value
-fLower = lambda value: value.lower()
-
-# time
-
-def unix_time_millis(dt):
-    epoch = datetime.datetime.utcfromtimestamp(0)
-    return (dt - epoch).total_seconds() * 1000.0
-
-def checkAccess(update):
-    return update.message.from_user.username in admins
-
-def isNewCommand(update):
-	print(update)
-	timenow = unix_time_millis(datetime.datetime.now())
-	messageTime = unix_time_millis(update.message.date)
-	dt = timenow - messageTime
-	print(update.message.text + " from " + update.message.from_user.username)
-	return dt < (20000)
-
-def isNewSeeds(update):
-    print(update.message.chat_id == newseeds)
-    print(update.message.chat_id)
-    return update.message.chat_id == newseeds
-
-# quotes
-
-def randomCunt():
-    return random.choice(["Ready Lets Go", "Here we go...", 
-        "Come to Daddy", "Oh boy, here I go killin' again!"])
-
-# session persistence
-
-def emptyConfig():
-    obj = dict()
-    obj['isPlaying'] = False
-    return obj
-
-def saveConfig(config):
-    with open(configFile, "wb") as fi:
-        pickle.dump(config, fi)
-        print("- {}".format(config))
-
-def loadConfig():
-    try:
-        with open(configFile, "rb") as fi:
-            config = pickle.load(fi)
-            return config
-    except:
-        print("config file is not found")
-        config = emptyConfig()
-        saveConfig(config)
-        return config
-
-# - commands
-
-# session
-
-def over(bot, update):
-    if not checkAccess(update):
-        return
-    config = loadConfig()
-    print(config)
-    if config['isPlaying'] == True:
-        if isNewCommand(update):
-            bot.sendMessage(newseeds, "#musictheatre it's OVER.")
-        endSession()
-    else:
-        update.message.reply_text("You betcha it is.")
-
-def abort(bot, update):
-    if not checkAccess(update):
-        return
-    config = loadConfig()
-    if config['isPlaying'] == True:
-        endSession()
-
-        if isNewCommand(update):
-            bot.sendMessage(newseeds, "#musictheatre it's ABORTED.")
-
-            # add 'aborted'
-            auth()
-            archiveCells = wks.range('F4:L4')
-            if archiveCells[4].value.encode('utf-8') == config['album']:
-                archiveCells[6].value = "aborted"
-                wks.update_cells(archiveCells)
-    else:
-        update.message.reply_text("I'll abort you, you fucking bitch.")
-
-def newAlbum(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-
-    message = update.message.text.split(" ", 1)[1].strip()
-
-    # new artist - album
-    if " - " in position:
-        artistName = position.split(" - ", 1)[0].strip()
-        albumName = position.split(" - ", 1)[1].strip()
-        newAlbumSet(bot, config, artistName, albumName, None, None)
-    # new {number}
-    else:
-        newAlbumSetPosition(bot, message)
-
-def newAlbumSetPosition(bot, position):
-    config = loadConfig()
-    if config['isPlaying'] == False:
-       if int(position) >= 4:
-            auth()
-            info = map(fValue, wks.range("B{0}:E{0}".format(position, position)))
-            newAlbumSet(bot, config, info[1], info[3], info[2], info[0])
-            
-            # archive as well
-            archiveDo(bot, position)
-    else:
-        bot.sendMessage(newseeds, "We're still in session.")
-
-def newAlbumSet(bot, config, artistName, albumName, year, suggested):
-    config['isPlaying'] = True
-    config['artist'] = artistName.encode('utf-8')
-    config['album'] = albumName.encode('utf-8')
-    config['track'] = ""
-
-    text = ""
-    if year is not None and suggested is not None:
-        config['year'] = year.encode('utf-8')
-        config['suggested'] = suggested.encode('utf-8')
-
-        text = "#musictheatre New Album: {0} - {1} ({2}) [Suggested by: {3}]".format(
-            config['artist'], config['album'], config['year'], config['suggested'])
-    else:
-        text = "#musictheatre New Album: {0} - {1}".format(
-            config['artist'], config['album'])
-
-    bot.sendMessage(newseeds, text)
-    saveConfig(config)
-
-def nextSong(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-    config = loadConfig()
-    if config['isPlaying'] == True:
-        trackName = update.message.text.split(" ", 1)[1].strip()
-        if len(trackName) > 0 and len(config['artist']) > 0 and trackName != config['track']:
-            config['track'] = trackName
-            text = "#musictheatre {0} - {1}".format(
-                config['artist'].encode('utf-8'), config['track'].encode('utf-8'))
-            bot.sendMessage(newseeds, text)
-            saveConfig(config)
-    else:
-        bot.sendMessage(newseeds, "What album was that again?")
-
-def endSession():
-    config = loadConfig()
-    config.pop('artist', None)
-    config.pop('track', None)
-    config.pop('album', None)
-    config.pop('suggested', None)
-    config.pop('year', None)
-    config['isPlaying'] = False
-    saveConfig(config)
-
-# current
-
-def currentAlbum(bot, update):
-    if not isNewCommand(update):
-        return
-    config = loadConfig()
-    if config['isPlaying'] == True:
-        if len(config['artist']) > 0 and len(config['album']) > 0:
-            text = "{0} by {1}".format(
-                config['album'].encode('utf-8'), config['artist'].encode('utf-8'))
-            if config['year'] is not None:
-                text += " ({})".format(config['year'].encode('utf-8'))
-            if config['suggested'] is not None:
-                text += " [Suggested by: {}]".format(
-                    config['suggested'].encode('utf-8'))
-            update.message.reply_text(text)
-    else:
-        update.message.reply_text("Nothing is playing.")
-
-
-def currentTrack(bot, update):
-    if not isNewCommand(update):
-        return
-    config = loadConfig()
-    if config['isPlaying'] == True:
-        if len(config['artist']) > 0 and len(config['album']) > 0 and len(config['track']) > 0:
-            text = "Now playing: {0} - {1} (from {2})".format(
-                config['artist'].encode('utf-8'), config['track'].encode('utf-8'), 
-                config['album'].encode('utf-8'))
-            update.message.reply_text(text)
-    else:
-        update.message.reply_text("Nothing is playing.")
-
-# countdown
-
-def processor_heavy_sleep(ms):  
-    start = time.clock()
-    end = start + ms / 1000.
-    while time.clock() < end:
-        continue
-    return start, time.clock()
-
-def cunt(bot, update):
-    if not isNewCommand(update):
-        return
-    if checkAccess(update):
-        message = [1]
-        if len(update.message.text.split(" ", 1)) > 1:
-            message = update.message.text.split(" ", 1)[1].strip()
-            print(message)
-        else:
-            message = randomCunt()
-        bot.sendMessage(newseeds, message.encode('utf-8'))
-        count = 5
-        while count:
-            bot.sendMessage(newseeds, "{}".format(count))
-            processor_heavy_sleep(999)
-            count -= 1
-        bot.sendMessage(newseeds, "PLAY!")
-
-        # also call /new
-        config = loadConfig()
-        if config['lastRoll'] is not None:
-            newAlbumSetPosition(bot, config['lastRoll'])
-
-# roll
-
-def getWeights(count):
-    values = list(range(0, count))
-    weights = list(map(lambda x: count - x/1.05, values))
-    weights = numpy.array(weights)
-    weights /= weights.sum()
-    return weights
-
-def getRandom(count):
-    values = list(range(0, count))
-    weights = getWeights(count)
-    return numpy.random.choice(values, p=weights)
-
-def roll(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-    config = loadConfig()
-    print(config)
-
-    if config['isPlaying'] == False:
-        auth()
-        suggestionNames = filter(fNonEmpty, map(fValue, wks.range('B4:B100')))
-        illegalNames = map(fLower, filter(fNonEmpty, map(fValue, wks.range('G4:G9'))))
-        suggestionsCount = len(suggestionNames)
-        print(illegalNames)
-        if suggestionsCount > 0:
-            for _ in range(5):
-
-                # get random (favor older suggestions)
-                result = getRandom(suggestionsCount-1)
-
-                spreadsheetNumber = result + 4
-                rolled = map(fValue, wks.range('A'+str(spreadsheetNumber)
-                    +':E'+ str(spreadsheetNumber)))
-                if not rolled[1].lower() in illegalNames:
-                    config['lastRoll'] = spreadsheetNumber
-                    saveConfig(config)
-
-                    bot.sendMessage(newseeds,
-                        "<b>Rolled {}</b>\n{} - {} ({})\nSuggested by: {}" .format(
-                            spreadsheetNumber, rolled[2].encode('utf-8'), 
-                            rolled[4].encode('utf-8'), rolled[3].encode('utf-8'), 
-                            rolled[1].encode('utf-8')), parse_mode="HTML")
-                    return
-                else:
-                    bot.sendMessage(newseeds,
-                        "Rolled {}. {} - illegal.".format(
-                            spreadsheetNumber, rolled[1].encode('utf-8')))
-        else:
-            update.message.reply_text("No suggestions found.")
-    else:
-    	update.message.reply_text("Another session is still on. I'm afraid I can't do that.")
-
-# archive
-
-def archive(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-    auth()
-    position = int(update.message.text.split(" ")[1])
-    archiveDo(bot, position)
-    
-def archiveDo(bot, position):
-    now = datetime.datetime.now()
-
-    archiveNames = filter(fNonEmpty, map(fValue, wks.range('G4:G1000')))
-    suggestionNames = filter(fNonEmpty, map(fValue, wks.range('B4:B100')))
-    lastSuggestion = len(suggestionNames) + 4
-    archiveLastPosition = len(archiveNames) + 4
-
-    rolledCells = wks.range('B'+str(position)+':E'+ str(position))
-    rolled = map(fValue, rolledCells)
-
-    # move archive cells down 1 row
-    archiveOldCells = wks.range('F4:L'+str(archiveLastPosition+1))
-    lenArchiveOld = len(archiveOldCells)
-    for i in reversed(range(len(archiveOldCells))):
-        if lenArchiveOld > i + 7:
-            archiveOldCells[i+7].value = archiveOldCells[i].value
-
-    wks.update_cells(archiveOldCells)
-
-    # move suggestion to F4
-    archiveCells = wks.range('F4:L4')
-    archiveCells[0].value = now.strftime("%d %b %y")
-    archiveCells[1].value = rolled[0]
-    archiveCells[2].value = rolled[1]
-    archiveCells[3].value = rolled[2]
-    archiveCells[4].value = rolled[3]
-    archiveCells[5].value = ""
-    archiveCells[6].value = ""
-
-    wks.update_cells(archiveCells)
-
-    # move suggestions up 1 row
-    suggestionCells = wks.range('B'+str(position)+':E'+str(lastSuggestion))
-    lenSuggestionCells = len(suggestionCells)
-    for i in range(len(suggestionCells)):
-        if lenSuggestionCells > i + 4:
-            suggestionCells[i].value = suggestionCells[i+4].value
-
-    wks.update_cells(suggestionCells)
-    bot.sendMessage(newseeds, "Suggestion moved to the archive.")
-
-# suggest
-
-def suggest(bot, update):
-    if not isNewCommand(update):
-        return
-    config = loadConfig()
-    if config['isPlaying'] == False:
-        bot.sendMessage(newseeds, "<b>Anyone in for a </b>#musictheatre<b> session?</b>", 
-            parse_mode="HTML")
-    else:
-        bot.sendMessage(newseeds, "Another session is still in place.")
-
-# help
-
-def start(bot, update):
-    update.message.reply_text(
-        "<b>Welcome to Music Theatre!</b>\nThis bot is designed to assist newseeds to discover music.\n", 
-        parse_mode="HTML")
-    help(bot, update)
-
-def help(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-    update.message.reply_text("Here's the list of commands:\n<b>/sheet</b> gives you the link to our spreadsheet\n<b>/tagme</b> to subscribe to session notifications [private message only]\n<b>/suggest</b> will ask if anyone wants to start a session\n<b>/admins</b> for the list of people who have admin access\n\nUse these while in session:\n<b>/song</b> or <b>/album</b> to find out what's playing".encode('utf-8'), 
-        parse_mode="HTML")
-
-def adminHelp(bot, update):
-    if not isNewCommand(update):
-        return
-    update.message.reply_text("Here's the list of admin commands:\n<b>/roll</b> to randomly pick a suggestion\n<b>/cunt</b> to initiate the countdown\n<b>/archive [roll]</b> to archive a suggestion\n<b>/new [roll]</b> will set current album playing and archive the suggestion\n\nUse these while in session:\n<b>/n</b> to set current song playing\n<b>/over or /abort</b> to end the session".encode('utf-8'), 
-        parse_mode="HTML")
-
-# say something
-
-def say(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-    if update.message.chat_id == newseeds:
-        return
-    message = update.message.text.split(" ", 1)[1].strip()
-    bot.sendMessage(newseeds, message.encode('utf-8'), parse_mode="HTML")
-
-def sticker(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-    if isNewSeeds(update):
-        return
-    message = update.message.text.split(" ", 1)[1].strip()
-    bot.sendSticker(newseeds, message)
-
-# admins
-
-def adminList(bot, update):
-    if not isNewCommand(update):
-        return
-    text = "These <i>(%username%)</i>s have access to the bot's #musictheatre session commands:\n"
-    for name in admins:
-        text += "- " + name + "\n"
-    text += "If they are not around, God help you."
-    update.message.reply_text(text.encode('utf-8'), parse_mode="HTML")
-
-# spreadshit link
-
-def shit(bot, update):
-    if not isNewCommand(update):
-        bot.sendMessage(newseeds, "Hey folks, our spreadshit is here: http://bit.ly/mtheatre", 
-            parse_mode="HTML")
-    else:
-        update.message.reply_text("Here's the link to our spreadshit: http://bit.ly/mtheatre")
-
-# tag
-
-def tagPeople(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-    if not isNewCommand(update):
-        return
-    config = loadConfig()
-    if 'tagList' in config and len(config['tagList']) > 0:
-        bot.sendMessage(newseeds, "Notifying {} people... (/taginfo to learn).".format(
-            len(config['tagList'])))
-        bot.sendMessage(newseeds, "<b>Anyone in for a </b>#musictheatre<b> session?</b>", 
-            parse_mode="HTML")
-        for id in config['tagList']:
-            if id != update.effective_user.id:
-                try:
-                    bot.sendMessage(id, "#musictheatre How about some music?")
-                except:
-                    print("{} blocked the bot.".format(id))
-    else:
-        bot.sendMessage(newseeds, "No one is subscribed for /tag updates.")
-
-def tagMe(bot, update):
-    if not isNewCommand(update):
-        return
-    if isNewSeeds(update):
-        update.message.reply_text(
-            "You have to private message me this command, because I am forbidden to initiate a chat with you.")
-    else:
-        id = update.effective_user.id
-        config = loadConfig()
-        if 'tagList' not in config:
-            config['tagList'] = []
-        if id not in config['tagList']:
-            config['tagList'].append(id)
-            bot.sendMessage(id, "You are now subscribed to /tag updates. Use /stop to unsubscribe.")
-        else:
-            bot.sendMessage(id, "You are already subscribed to /tag updates.")
-        saveConfig(config)
-
-def dontTagMe(bot, update):
-    id = update.effective_user.id
-    config = loadConfig()
-    if 'tagList' not in config and id not in config['tagList']:
-        config['tagList'] = []
-        bot.sendMessage(id, "You are not subscribed to /tag updates.")
-    else:
-        config['tagList'].remove(id)
-        bot.sendMessage(id, "You are now unsubscribed from /tag updates.")
-    saveConfig(config)
-
-def taginfo(bot, update):
-    update.message.reply_text("PM /tagme to the bot to subscribe to #musictheatre updates.")
-
-# retarded
-
-def slow(bot, update):
-	bot.sendSticker(newseeds, retardStickerId)
-
-
-# manage suggestions
-
-def countSuggestions(bot, update):
-    if not isNewCommand(update):
-        return
-
-    array = update.message.text.split(" ")
-    if len(array) < 2:
-        update.message.reply_text("Enter the name to see how many albums he has suggested.")
-        return
-
-    name = update.message.text.split(" ")[1]
-    auth()
-
-    originalRows = map(fValue, wks.range('B4:B100'))
-    foundRows = filter(lambda value: value.lower() == name.lower(), originalRows)
-    suggestionsCount = len(foundRows)
-
-    originalArchive = map(fValue, wks.range('G4:G1000'))
-    foundArchive = filter(lambda value: value.lower() == name.lower(), originalArchive)
-    archiveCount = len(foundArchive)
-
-    if suggestionsCount > 0:
-        if archiveCount > 0:
-            update.message.reply_text("{} current suggestions by {}. Also {} in archive.".format(
-                suggestionsCount, foundRows[0], archiveCount))
-        else :
-            update.message.reply_text("{} current suggestions by {}.".format(
-                suggestionsCount, foundRows[0]))
-    elif archiveCount > 0:
-        update.message.reply_text("No current suggestion by {} but there are {} in archive.".format(
-            foundArchive[0]))
-    else:
-        update.message.reply_text("No such thing.")
-
-def addSuggestion(bot, update):
-    if not isNewCommand(update):
-        return
-    if not checkAccess(update):
-        return
-
-    array = re.split("[;\n]", update.message.text)
-
-    print(array)
-    if len(array) != 4:
-        update.message.reply_text("Enter suggester name, artist, year and title of release (use semicolon or new line to divide input)\nLike this: \"Yaro; Pink Floyd; The Dark Side of the Moon; 1973\"")
-        return
-
-    auth()
-    
-    # name artist year album
-    name = array[0].strip()[5:]
-    artist = array[1].strip()
-    album = array[2].strip()
-    year = int(array[3].strip())
-
-    if year > 2019:
-        update.message.reply_text("What kind of future bullshit is that?")
-        return
-        
-    if year < 999:
-        update.message.reply_text("Music of the generation before smartphones? Nah!")
-        return
-
-    if year < 1900:
-        update.message.reply_text("Let's hear the cavemen music...")
-
-    suggestionNames = filter(fNonEmpty, map(fValue, wks.range('B4:B100')))
-    foundRows = filter(lambda value: value.lower() == name.lower(), suggestionNames)
-
-    if len(foundRows) >= 3:
-        update.message.reply_text("This one already has enough 3 suggestions.")
-        return
-
-    newSuggestion = len(suggestionNames) + 4
-
-    # add suggestion
-    newCells = wks.range('B'+str(newSuggestion)+':E'+str(newSuggestion))
-    newCells[0].value = name
-    newCells[1].value = artist
-    newCells[2].value = year
-    newCells[3].value = album
-
-    wks.update_cells(newCells)
-
-    update.message.reply_text("{} by {} ({}) is now suggested by {}.".format(
-        album.encode('utf-8'), artist.encode('utf-8'), year, name.encode('utf-8')))
-
-
-def rollInfo(bot, update):
-    if not isNewCommand(update):
-        return
-
-    auth()
-
-    suggestions = filter(fNonEmpty, map(fValue, wks.range('B4:B100')))
-    count = len(suggestions)
-
-    if count < 2:
-        return
-
-    values = getWeights(count)
-
-    array = update.message.text.split(" ")
-    if len(array) > 1:
-        try: 
-            position = int(array[1])
-            prob = values[position - 4] * 100
-            update.message.reply_text("Probability of rolling {0} is {1:.1f}%".format(position, prob))
-            return
-        except Exception as e: print(e)
-
-    first = values[0] * 100
-    last = values[-1] * 100
-    update.message.reply_text("Roll probability (linear distribution):\nOldest: {0:.1f}%\nNewest: {1:.1f}%".format(first, last))
-
-# work
-
-def test(bot, update):
-    if not isNewCommand(update):
-        return
-    if update.message.from_user.username == "ysoftware":
-        bot.sendMessage(101193334, loadConfig())
-
-def error_callback(bot, update, error):
-    print(error)
-
-# commands
+# setup bot
 
 logging.basicConfig(level=logging.WARN, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 
@@ -661,51 +25,52 @@ REQUEST_KWARGS={
 updater = Updater(TOKEN, request_kwargs=REQUEST_KWARGS)
 
 # sheet
-updater.dispatcher.add_handler(CommandHandler('shit', shit))
-updater.dispatcher.add_handler(CommandHandler('sheet', shit))
-updater.dispatcher.add_handler(CommandHandler('spreadshit', shit))
-updater.dispatcher.add_handler(CommandHandler('spreadsheet', shit))
+updater.dispatcher.add_handler(CommandHandler('shit', other_commands.shit))
+updater.dispatcher.add_handler(CommandHandler('sheet', other_commands.shit))
+updater.dispatcher.add_handler(CommandHandler('spreadshit', other_commands.shit))
+updater.dispatcher.add_handler(CommandHandler('spreadsheet', other_commands.shit))
 
 # user session
-updater.dispatcher.add_handler(CommandHandler('suggest', suggest))
-updater.dispatcher.add_handler(CommandHandler('song', currentTrack))
-updater.dispatcher.add_handler(CommandHandler('album', currentAlbum))
-updater.dispatcher.add_handler(CommandHandler('tag', tagPeople))
+updater.dispatcher.add_handler(CommandHandler('suggest', start_commands.suggest))
+updater.dispatcher.add_handler(CommandHandler('song', session_commands.currentTrack))
+updater.dispatcher.add_handler(CommandHandler('album', session_commands.currentAlbum))
+updater.dispatcher.add_handler(CommandHandler('tag', other_commands.tagPeople))
 
 # session notifications
-updater.dispatcher.add_handler(CommandHandler('tagme', tagMe))
-updater.dispatcher.add_handler(CommandHandler('stop', dontTagMe))
-updater.dispatcher.add_handler(CommandHandler('fuckoff', dontTagMe))
-updater.dispatcher.add_handler(CommandHandler('unsub', dontTagMe))
-updater.dispatcher.add_handler(CommandHandler('taginfo', taginfo))
+updater.dispatcher.add_handler(CommandHandler('tagme', other_commands.tagMe))
+updater.dispatcher.add_handler(CommandHandler('stop', other_commands.dontTagMe))
+updater.dispatcher.add_handler(CommandHandler('fuckoff', other_commands.dontTagMe))
+updater.dispatcher.add_handler(CommandHandler('unsub', other_commands.dontTagMe))
+updater.dispatcher.add_handler(CommandHandler('taginfo', other_commands.taginfo))
 
-# admin 
-updater.dispatcher.add_handler(CommandHandler('test', test))
-updater.dispatcher.add_handler(CommandHandler('s', sticker))
-updater.dispatcher.add_handler(CommandHandler('b', say))
+# # admin 
+updater.dispatcher.add_handler(CommandHandler('test', other_commands.test))
+updater.dispatcher.add_handler(CommandHandler('s', other_commands.sticker))
+updater.dispatcher.add_handler(CommandHandler('b', other_commands.say))
+updater.dispatcher.add_handler(CommandHandler('removeRoll', other_commands.removeRoll))
 
-# admin session 
-updater.dispatcher.add_handler(CommandHandler('roll', roll))
-updater.dispatcher.add_handler(CommandHandler('archive', archive))
-updater.dispatcher.add_handler(CommandHandler('cunt', cunt))
-updater.dispatcher.add_handler(CommandHandler('new', newAlbum))
-updater.dispatcher.add_handler(CommandHandler('n', nextSong))
-updater.dispatcher.add_handler(CommandHandler('abort', abort))
-updater.dispatcher.add_handler(CommandHandler('over', over))
+# # admin session 
+updater.dispatcher.add_handler(CommandHandler('roll', start_commands.roll))
+updater.dispatcher.add_handler(CommandHandler('cunt', start_commands.cunt))
+updater.dispatcher.add_handler(CommandHandler('archive', sheet_commands.archive))
+updater.dispatcher.add_handler(CommandHandler('new', session_commands.newAlbum))
+updater.dispatcher.add_handler(CommandHandler('n', session_commands.nextSong))
+updater.dispatcher.add_handler(CommandHandler('abort', session_commands.abort))
+updater.dispatcher.add_handler(CommandHandler('over', session_commands.over))
 
 # other 
-updater.dispatcher.add_handler(CommandHandler('slow', slow))
-updater.dispatcher.add_handler(CommandHandler('start', start))
-updater.dispatcher.add_handler(CommandHandler('help', help))
-updater.dispatcher.add_handler(CommandHandler('adminHelp', adminHelp))
-updater.dispatcher.add_handler(CommandHandler('admins', adminList))
+updater.dispatcher.add_handler(CommandHandler('slow', other_commands.slow))
+updater.dispatcher.add_handler(CommandHandler('start', other_commands.start))
+updater.dispatcher.add_handler(CommandHandler('help', other_commands.help))
+updater.dispatcher.add_handler(CommandHandler('adminHelp', other_commands.adminHelp))
+updater.dispatcher.add_handler(CommandHandler('admins', other_commands.adminList))
 
 # manage suggestions
-updater.dispatcher.add_handler(CommandHandler('rollinfo', rollInfo))
-updater.dispatcher.add_handler(CommandHandler('count', countSuggestions))
-updater.dispatcher.add_handler(CommandHandler('add', addSuggestion))
+updater.dispatcher.add_handler(CommandHandler('rollinfo', sheet_commands.rollInfo))
+updater.dispatcher.add_handler(CommandHandler('count', sheet_commands.countSuggestions))
+updater.dispatcher.add_handler(CommandHandler('add', sheet_commands.addSuggestion))
 
-updater.dispatcher.add_error_handler(error_callback)
+updater.dispatcher.add_error_handler(other_commands.error_callback)
 
 updater.start_polling()
 updater.idle()
