+++ utility.py
@@ -1,0 +1,5 @@
+# utilities
+
+fValue = lambda cell: cell.value
+fNonEmpty = lambda value: value
+fLower = lambda value: value.lower()
